+++ flake.nix
@@ -67,10 +67,11 @@
         {
           default = pkgs.mkShell {
             packages = with pkgs; [
-              vicinae.packages.${system}.default
+              # vicinae.packages.${system}.default
               self.formatter.${system}
               nodejs
               typescript-language-server
+              nixd
             ];
           };
         }