+++ extensions/bluetooth/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added a changelog
 - Added a discoverable command
-- added a preference to make the connect action toggleable instead of haveing 2 commands for connect and disconnect
+- added a preference to make the connect action toggleable instead of having 2 commands for connect and disconnect
 - Big refactor of the codebase
     - reworked the parsing of bluetoothctl output to add convert to typed lines